% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{bbox_to_polygon}
\alias{bbox_to_polygon}
\title{Convert bbox to polygon}
\usage{
bbox_to_polygon(bbox, crs)
}
\arguments{
\item{bbox}{Bounding box (sf, sfc, or named vector)}

\item{crs}{CRS to use}
}
\value{
sfc polygon
}
\description{
Convert bbox to polygon
}
\keyword{internal}
