% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{determine_sides}
\alias{determine_sides}
\title{Determine which side of a line each point is on}
\usage{
determine_sides(points, line)
}
\arguments{
\item{points}{Points to classify}

\item{line}{Line geometry}
}
\value{
Vector of side assignments (1 or 2)
}
\description{
Determine which side of a line each point is on
}
\keyword{internal}
