% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{find_border_width}
\alias{find_border_width}
\title{Find buffer width that produces target area per side}
\usage{
find_border_width(border_geom, target_area, mask_geom, crs)
}
\arguments{
\item{border_geom}{Border line geometry (sfc)}

\item{target_area}{Target area for each side}

\item{mask_geom}{Optional mask geometry}

\item{crs}{CRS of the geometries}
}
\value{
Buffer width
}
\description{
Find buffer width that produces target area per side
}
\keyword{internal}
