% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{find_halo_width}
\alias{find_halo_width}
\title{Find additional halo width for target halo area}
\usage{
find_halo_width(border_geom, core_width, target_halo_area, mask_geom, crs)
}
\arguments{
\item{border_geom}{Border line geometry}

\item{core_width}{Width of core zone}

\item{target_halo_area}{Target area for halo per side}

\item{mask_geom}{Optional mask geometry}

\item{crs}{CRS}
}
\value{
Additional width for halo
}
\description{
Find additional halo width for target halo area
}
\keyword{internal}
