% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{plot.aoe_border_result}
\alias{plot.aoe_border_result}
\title{Plot method for aoe_border_result}
\usage{
\method{plot}{aoe_border_result}(x, ...)
}
\arguments{
\item{x}{An aoe_border_result object}

\item{...}{Additional arguments passed to plot}
}
\value{
NULL (called for side effect)
}
\description{
Plot method for aoe_border_result
}
