% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_class.R
\name{print.aoe_result}
\alias{print.aoe_result}
\title{Print method for aoe_result}
\usage{
\method{print}{aoe_result}(x, ...)
}
\arguments{
\item{x}{An aoe_result object}

\item{...}{Additional arguments passed to print.sf}
}
\value{
Invisibly returns x
}
\description{
Print method for aoe_result
}
