% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{split_by_buffer}
\alias{split_by_buffer}
\title{Split polygon using buffer method (fallback)}
\usage{
split_by_buffer(polygon, line, crs)
}
\arguments{
\item{polygon}{Polygon to split}

\item{line}{Line to split by}

\item{crs}{CRS}
}
\value{
List with side1 and side2
}
\description{
Split polygon using buffer method (fallback)
}
\keyword{internal}
