% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_border.R
\name{split_by_line}
\alias{split_by_line}
\title{Split a polygon by a line into two sides}
\usage{
split_by_line(polygon, line, crs)
}
\arguments{
\item{polygon}{Polygon geometry to split}

\item{line}{Line geometry to split by}

\item{crs}{CRS}
}
\value{
List with side1 and side2 geometries
}
\description{
Split a polygon by a line into two sides
}
\keyword{internal}
