% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_class.R
\name{[.aoe_result}
\alias{[.aoe_result}
\title{Subset method for aoe_result}
\usage{
\method{[}{aoe_result}(x, i, ...)
}
\arguments{
\item{x}{An aoe_result object}

\item{i}{Row indices}

\item{...}{Additional arguments passed to sf subsetting}
}
\value{
An aoe_result object (or sf if support_id removed)
}
\description{
Preserves aoe_result attributes when subsetting.
}
