% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoe_class.R
\name{summary.aoe_result}
\alias{summary.aoe_result}
\title{Summary method for aoe_result}
\usage{
\method{summary}{aoe_result}(object, ...)
}
\arguments{
\item{object}{An aoe_result object}

\item{...}{Additional arguments (ignored)}
}
\value{
An aoe_summary_result object
}
\description{
Summary method for aoe_result
}
